% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAllElevations.R
\name{downloadAllElevations}
\alias{downloadAllElevations}
\title{Download elevations for each country}
\usage{
downloadAllElevations()
}
\arguments{
\item{NULL}{Input values}
}
\value{
NULL No values output
}
\description{
Download elevations for each country
}
\examples{
downloadAllElevations()
}
