% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAllShapes.R
\name{downloadAllShapes}
\alias{downloadAllShapes}
\title{Download shape file for all countries}
\usage{
downloadAllShapes()
}
\arguments{
\item{NULL}{Input values}
}
\value{
NULL No values output
}
\description{
Download shape file for all countries
}
\examples{
downloadAllShapes()
}
