% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCapital.R
\name{getCapital}
\alias{getCapital}
\title{Relate country with capital}
\usage{
getCapital(selectGeneral)
}
\arguments{
\item{selectGeneral}{data for main map}
}
\value{
result Dataframe adding capital, latitud, longitud of a coutry to plot in main screen
}
\description{
Relate country with capital
}
\examples{
getCapital(selectGeneral)
}
