% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIso3.R
\name{getIso3}
\alias{getIso3}
\title{Relate ISO3 code with country}
\usage{
getIso3(data)
}
\arguments{
\item{data}{data for main map}
}
\value{
result Dataframe adding ISO3 code for that country
}
\description{
Relate ISO3 code with country
}
\examples{
getCapital(selectGeneral)
}
