% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Download all assemblages values}
\usage{
loadData()
}
\arguments{
\item{NULL}{No values input}
}
\value{
list(taxon = taxon, assemblages = assemblages, index = index) No values output
}
\description{
Download all assemblages values
}
\examples{
loadData()
}
