% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\name{plotMap}
\alias{plotMap}
\title{Plot the map}
\usage{
plotMap(input, data)
}
\arguments{
\item{input}{Input values}

\item{data}{Data to plot}
}
\value{
NULL No values output
}
\description{
Plot the map
}
\examples{
plotMap(input,data)
}
