% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMapForCountry.R
\name{plotMapForCountry}
\alias{plotMapForCountry}
\title{Plot the map}
\usage{
plotMapForCountry(input, data)
}
\arguments{
\item{input}{Input values}

\item{data}{Data to plot}
}
\value{
NULL No values output
}
\description{
Plot the map
}
\examples{
plotMapForCountry(input,data)
}
