% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCountry.R
\name{selectCountry}
\alias{selectCountry}
\title{Select country data}
\usage{
selectCountry(selectGeneral, input_country)
}
\arguments{
\item{selectGeneral}{#Datastream with indexed values}

\item{input_country}{#Country selected}
}
\value{
data for this country selected
}
\description{
Select country data
}
\examples{
selectCountry(selectGeneral,country)
}
