% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGeneral.R
\name{selectGeneral}
\alias{selectGeneral}
\title{Select general data}
\usage{
selectGeneral(index, assemblages, taxon)
}
\arguments{
\item{index}{#Datastream index}

\item{assemblages}{#Datastream assemblages}

\item{taxon}{#Datastream taxon}
}
\value{
unique_merged_data Result of query
}
\description{
Select general data
}
\examples{
selectGeneral(index,assemblages,taxon)
}
